+++ Source/DTrackInput/Private/DTrackInputFeatures.h
@@ -1,0 +1,24 @@
+
+#pragma once
+
+#include "InputCoreTypes.h"
+
+
+struct FDTrackInputKey
+{
+	static const FKey Flystick_Trigger;
+	static const FKey Flystick_Button1;
+	static const FKey Flystick_Button2;
+	static const FKey Flystick_Button3;
+	static const FKey Flystick_Button4;
+	static const FKey Flystick_Button5;
+	static const FKey Flystick_Button6;
+	static const FKey Flystick_Button7;
+	static const FKey Flystick_Button8;
+
+	static const FKey Flystick_JoystickX;
+	static const FKey Flystick_JoystickY;
+};
+
+
+